Code improvements from PR comments
